requirements.txt
highrise-bot-sdk>=24.1.0
python-dotenv>=1.0.1
aiohttp>=3.9.4